+++ packages/components/src/common/deviceDetection/index.js
@@ -1,0 +1,3 @@
+import isIosDevice from './deviceDetection';
+
+export default isIosDevice;